move CpuBuffersInfo to base class
