+++ onnxruntime/core/providers/cuda/cuda_stream_handle.h
@@ -5,7 +5,6 @@
 #include "core/providers/cuda/shared_inc/cuda_utils.h"
 #include "core/providers/cuda/shared_inc/cuda_call.h"
 #include "core/framework/stream_handles.h"
-#include "core/framework/allocator.h"
 
 namespace onnxruntime {
