+++ onnxruntime/core/providers/rocm/rocm_stream_handle.h
@@ -3,7 +3,6 @@
 //#include "core/providers/cuda/shared_inc/cuda_utils.h"
 #include "core/providers/rocm/shared_inc/rocm_call.h"
 #include "core/framework/stream_handles.h"
-#include "core/framework/allocator.h"
 
 namespace onnxruntime {
